Add files via upload
This is a script to read a file with a list of host, read a list of filename from a firectory and with specific extension, and upload the files to all host with an ftp username, password and port. If very useful to replicate software upgrade to many servers at the same time.